fix login button on new login page to work with DynamoDB instead of ceramic
