+++ components/HeaderNav/HeaderNav.tsx
@@ -1,71 +1,8 @@
 import React, { useState } from "react";
 import { Flex, HStack, Text, Img } from "@chakra-ui/react";
 
-import Link from "next/link";
-import { useRouter } from "next/router";
-import Web3 from "web3";
-import Web3Modal from "web3modal";
-import WalletConnectProvider from "@walletconnect/web3-provider";
-import { UserData } from "../../pages/user";
 const HeaderNav = (props) => {
   const whichPage = props.whichPage;
-  const [show, setShow] = useState(false);
-  const toggleMenu = () => setShow(!show);
-
-  const [isSubmitted, setIsSubmitted] = React.useState(false);
-  const [loaded, setLoaded] = useState(false);
-  const router = useRouter();
-
-  const getUserByWallet = async (userWallet) => {
-    let lowerCaseWallet = userWallet.toLowerCase();
-    const res = await fetch(`/api/users/wallet/${lowerCaseWallet}`);
-
-    const data: any = await res.json();
-
-    console.log("RETRIEVE USER BY WALLET YO");
-    return data;
-  };
-
-  const handleWalletConnect = async () => {
-    const web3Modal = new Web3Modal({
-      disableInjectedProvider: false,
-      network: "rinkeby",
-      cacheProvider: false,
-      providerOptions: {
-        walletconnect: {
-          package: WalletConnectProvider,
-          options: {
-            rpc: {
-              1: "https://eth-rinkeby.alchemyapi.io/v2/QtLM8rW9nB6DobDu8KQx-7fYMS2rBlky",
-            },
-          },
-        },
-      },
-    });
-    const provider = await web3Modal.connect();
-    const web3 = new Web3(provider);
-    const accounts = await web3.eth.getAccounts();
-    const currAccount = accounts[0];
-    console.log(currAccount);
-
-    setIsSubmitted(true);
-    try {
-      let userData: UserData = await getUserByWallet(currAccount);
-      console.log(userData);
-
-      if (userData && userData.user_name && userData.user_wallet) {
-        router.push(`/${userData.user_name}`);
-        setIsSubmitted(false);
-      } else {
-        console.log("No profile, pls create one...");
-        router.push("/steps");
-      }
-    } catch (err) {
-      console.log("error: ", err);
-      router.push("/steps");
-      setLoaded(true);
-    }
-  };
 
   return (
     <Flex
