+++ data/dynamodb.js
@@ -10,14 +10,14 @@ const getDynamoDBClient = () => {
     : "us-east-2";
 
   // Only needed with local development.
-  // if (process.env.LOCAL_DYNAMO_DB_ENDPOINT) {
-  //   AWS.config.update({
-  //     accessKeyId: "xxxx",
-  //     secretAccessKey: "xxxx",
-  //     region: "us-east-1",
-  //     endpoint: "http://localhost:8000",
-  //   });
-  // }
+  if (process.env.LOCAL_DYNAMO_DB_ENDPOINT) {
+    AWS.config.update({
+      accessKeyId: "xxxx",
+      secretAccessKey: "xxxx",
+      region: "us-east-1",
+      endpoint: "http://localhost:8000",
+    });
+  }
 
   const options = {
     convertEmptyValues: true,
